Домашнее задание 4. Созданы и импортированы модули
